fix webgpu sample assets

+++ sources/indexed-depth.wgsl
@@ -44,7 +44,7 @@ fn depth_preview_vs(@builtin(vertex_index) vertexId: u32) -> PreviewOut {
 
 @fragment
 fn depth_preview_fs(input: PreviewOut) -> @location(0) vec4<f32> {
-    let r36: vec2<f32> = (vec2<f32>(f32(i32(textureDimensions(usl_g0_b0, u32(0)).x)), f32(i32(textureDimensions(usl_g0_b0, u32(0)).y))) - vec2<f32>(1.0));
+    let r36: vec2<f32> = (vec2<f32>(f32(i32(textureDimensions(usl_g0_b0, 0u).x)), f32(i32(textureDimensions(usl_g0_b0, 0u).y))) - vec2<f32>(1.0));
     let r37: vec2<f32> = (input.uv * r36);
     let r38: vec2<u32> = vec2<u32>(r37);
     let r39: f32 = textureLoad(usl_g0_b0, vec2<i32>(r38), i32(0));
